requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
pandas==2.2.2
pyarrow==17.0.0
python-dotenv==1.0.1
tenacity==8.5.0